+++ Diego-Lofiego/Modelo/ProductoVendido.cs
@@ -2,15 +2,15 @@
 {
     class ProductoVendido : IId
     {
-        private int _id;
-        private int _idProducto;
-        private long _stock;
-        private int _idVenta;
+        private long _id;
+        private long _idProducto;
+        private int _stock;
+        private long _idVenta;
 
-        public int Id { get { return _id; } set { _id = value; } }
-        public int IdProducto { get { return _idProducto; } set { _idProducto = value; } }
-        public long Stock { get { return _stock; } set { _stock = value; } }
-        public int IdVenta { get { return _idVenta; } set { _idVenta = value; } }
+        public long Id { get { return _id; } set { _id = value; } }
+        public long IdProducto { get { return _idProducto; } set { _idProducto = value; } }
+        public int Stock { get { return _stock; } set { _stock = value; } }
+        public long IdVenta { get { return _idVenta; } set { _idVenta = value; } }
 
         public ProductoVendido()
         {
@@ -20,7 +20,7 @@
             _idVenta = 0;
         }
 
-        public ProductoVendido(int id, int idProducto, long stock, int idVenta)
+        public ProductoVendido(long id, long idProducto, int stock, long idVenta)
         {
             _id = id;
             _idProducto = idProducto;
